Switching the weighting type from int to double. This was done to allow weights to represent probabilities (ie with total weight = 1). This was useful in the Jack's Car Rental problem as the probabilities were already computed at the beginning.

+++ src/rl/Environment.h
@@ -14,7 +14,7 @@ namespace rl {
  * https://google.github.io/styleguide/cppguide.html#Integer_Types
  */
 using ID = int;
-using Weight = int;
+using Weight = double;
 
 
 class State {
